Registro con email manual funcional

+++ backend/Seminario/src/main/java/Grupo11/Seminario/Entities/Telefono.java
@@ -23,6 +23,6 @@ public class Telefono {
     private Integer codigo;
 
     @Column(nullable = false, name = "numero")
-    private Integer numero;
+    private Long numero;
 
 }
